JK - Fixed possible x86 floating point problem in glb_hybrid_bracket

+++ Distribution/source/glb_min_sup.c
@@ -503,8 +503,8 @@ int glb_hybrid_bracket(double *P, double *xi, double *a, double *b, double *c, i
                        glb_minimize_func f, int new_rates_flag, void *user_data)
 {
   double ax, bx, cx;
-  double fa, fb, fc, fu;
-  double u, ulim, q, r;
+  long double fa, fb, fc, fu; /* long double declaration avoids problems with  */
+  double u, ulim, q, r;       /* extra floating point digits on x86 processors */
   int i, j;
 
   ax = 0.0;
